ユーザーID列の型を`Long`に変更し、`NoSingleRecordErr`例外や関連ロジックを実装。`DbPersister`におけるユーザー取得処理を拡張。

+++ ktse/src/main/kotlin/net/kigawa/keruta/ktse/persist/db/table/UserTable.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object UserTable: Table("user") {
-    val id = integer("id").autoIncrement()
+    val id = long("id").autoIncrement()
     val createdAt = datetime("created_at")
     override val primaryKey = PrimaryKey(id)
 }
